finish generalizing all processing to all .md files in functions/ variables/ interfaces/
